Entirely rework the dropzone recalculations. The previous technique was flawed in multiple ways.
Current implementation only handles vertical scrolling. Presumably if this passes muster horizontal
scrolling can be accounted for as well.

The current technique attempts to determine if the element is partially obscured, in which case
special allocations are made for calculating it's offsets and proportions.

If the element is fully exposed then only the offset is calculated, as the proportion will be unchanged.

Note: Elements that are completely obscured are not properly handled. The proportions of such elements
seemingly will end up becoming negative due to the partially obscured branches, preventing these
zones from accepting a droppable when they are off the screen.