Rework the dropzone calculations again
Rather than try to take all the various offsets into account, leverage
getBoundingClientRect(), which has pretty good cross browser support.

This method returns back the bounding rectangle of any specified element.
Using it we can calculate the effective viewport that the droppable is
displayed in, and then determine how best to handle the droppable itself.

Ultimately it is likely the above approach can be extended to determine
whether an element is even viewable within the effective viewport. If not
we can handle it appropriately.

The technique described seemingly is working as expected for all cases
except:
- horizontal scrolling (I have no tests yet)
- top window scrolling (a regression that hopefully is easy to fix)