Extend dropzone calculations
The previous iteration only accounted for the top/bottom of a dropzone, added
handling for left/right as well.

While here I also reworked some of the logic. Rather than handling the normal
unobscured case if no changes were made to the top/bottom, we now handle the
basic case and then handle the edge cases if such handling is necessary, cuts
down some of the lines and removes spurious comparisons.